+++ src/account/update.rs
@@ -3,7 +3,10 @@
 
 use std::collections::HashMap;
 
-use iota_client::{api_types::response::OutputMetadataResponse, block::output::OutputId, Client};
+use iota_client::{
+    block::output::{dto::OutputMetadataDto, OutputId},
+    Client,
+};
 
 use crate::account::{
     handle::AccountHandle,
@@ -34,7 +37,7 @@ impl AccountHandle {
         &self,
         addresses_with_unspent_outputs: Vec<AddressWithUnspentOutputs>,
         unspent_outputs: Vec<OutputData>,
-        spent_or_unsynced_output_metadata_map: HashMap<OutputId, Option<OutputMetadataResponse>>,
+        spent_or_unsynced_output_metadata_map: HashMap<OutputId, Option<OutputMetadataDto>>,
         options: &SyncOptions,
     ) -> crate::Result<()> {
         log::debug!("[SYNC] Update account with new synced transactions");
@@ -265,6 +268,7 @@ impl AccountHandle {
     }
 
     // Should only be called from the AccountManager so all accounts are on the same state
+    // Will update the addresses with a possible new Bech32 HRP and clear the inaccessible_incoming_transactions.
     pub(crate) async fn update_account_with_new_client(&mut self, client: Client) -> crate::Result<()> {
         self.client = client;
         let bech32_hrp = self.client.get_bech32_hrp().await?;
@@ -279,6 +283,18 @@ impl AccountHandle {
         for address in &mut account.internal_addresses {
             address.address.bech32_hrp = bech32_hrp.clone();
         }
+
+        account.inaccessible_incoming_transactions.clear();
+
+        #[cfg(feature = "storage")]
+        {
+            log::debug!(
+                "[SYNC] storing account {} after updating it with new client options",
+                account.alias()
+            );
+            self.save(Some(&account)).await?;
+        }
+
         Ok(())
     }
 }