+++ src/account_manager/builder.rs
@@ -14,15 +14,16 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::Mutex;
 use tokio::sync::RwLock;
 
-#[cfg(feature = "storage")]
-use crate::account::handle::AccountHandle;
 #[cfg(feature = "events")]
 use crate::events::EventEmitter;
+#[cfg(all(feature = "storage", not(feature = "rocksdb")))]
+use crate::storage::adapter::memory::Memory;
 #[cfg(feature = "storage")]
-use crate::storage::constants::ROCKSDB_FOLDERNAME;
-#[cfg(feature = "storage")]
-use crate::storage::manager::ManagerStorage;
-use crate::{account_manager::AccountManager, ClientOptions};
+use crate::{
+    account::handle::AccountHandle,
+    storage::{constants::default_storage_path, manager::ManagerStorage},
+};
+use crate::{account_manager::AccountManager, ClientOptions, Error};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 /// Builder for the account manager.
@@ -48,10 +49,10 @@ pub struct StorageOptions {
 impl Default for StorageOptions {
     fn default() -> Self {
         StorageOptions {
-            storage_path: ROCKSDB_FOLDERNAME.into(),
+            storage_path: default_storage_path().into(),
             storage_file_name: None,
             storage_encryption_key: None,
-            manager_store: ManagerStorage::Rocksdb,
+            manager_store: ManagerStorage::default(),
         }
     }
 }
@@ -121,99 +122,105 @@ impl AccountManagerBuilder {
                 return Err(crate::Error::MissingParameter("secret_manager"));
             }
         }
-        #[cfg(feature = "storage")]
+        #[cfg(all(feature = "rocksdb", feature = "storage"))]
         let storage =
             crate::storage::adapter::rocksdb::RocksdbStorageAdapter::new(storage_options.storage_path.clone())?;
+        #[cfg(all(not(feature = "rocksdb"), feature = "storage"))]
+        let storage = Memory::default();
+
         #[cfg(feature = "storage")]
         let storage_manager = crate::storage::manager::new_storage_manager(
             None,
             Box::new(storage) as Box<dyn crate::storage::adapter::StorageAdapter + Send + Sync>,
         )
         .await?;
+
         #[cfg(feature = "storage")]
-        {
-            let manager_builder = storage_manager.lock().await.get_account_manager_data().await.ok();
-            let (client_options, secret_manager, coin_type) = match manager_builder {
-                Some(data) => {
-                    // prioritise provided client_options and secret_manager over stored ones
-                    let client_options = match self.client_options {
-                        Some(options) => options,
-                        None => data
-                            .client_options
-                            .ok_or(crate::Error::MissingParameter("client_options"))?,
-                    };
-                    let secret_manager = match self.secret_manager {
-                        Some(secret_manager) => secret_manager,
-                        None => data
-                            .secret_manager
-                            .ok_or(crate::Error::MissingParameter("secret_manager"))?,
-                    };
-                    let coin_type = match self.coin_type {
-                        Some(coin_type) => coin_type,
-                        None => data
-                            .coin_type
-                            .ok_or(crate::Error::MissingParameter("coin_type (IOTA: 4218, Shimmer: 4219)"))?,
-                    };
-                    (client_options, secret_manager, coin_type)
-                }
-                // If no account manager data exist, we will set it
-                None => {
-                    // Store account manager data in storage
-                    storage_manager.lock().await.save_account_manager_data(&self).await?;
-                    (
-                        self.client_options
-                            .ok_or(crate::Error::MissingParameter("client_options"))?,
-                        self.secret_manager
-                            .ok_or(crate::Error::MissingParameter("secret_manager"))?,
-                        self.coin_type
-                            .ok_or(crate::Error::MissingParameter("coin_type (IOTA: 4218, Shimmer: 4219)"))?,
-                    )
-                }
-            };
+        let read_manager_builder = match storage_manager.lock().await.get_account_manager_data().await {
+            Ok(x) => Some(x),
+            Err(Error::RecordNotFound(_)) => None,
+            Err(e) => return Err(e),
+        };
+        #[cfg(not(feature = "storage"))]
+        let read_manager_builder: Option<AccountManagerBuilder> = None;
 
-            let client = client_options.clone().finish()?;
+        // prioritise provided client_options and secret_manager over stored ones
+        let new_provided_client_options = if self.client_options.is_none() {
+            let loaded_client_options = read_manager_builder
+                .as_ref()
+                .and_then(|data| data.client_options.clone())
+                .ok_or(crate::Error::MissingParameter("client_options"))?;
 
-            let accounts = storage_manager.lock().await.get_accounts().await.unwrap_or_default();
+            // Update self so it gets used and stored again
+            self.client_options.replace(loaded_client_options);
+            false
+        } else {
+            true
+        };
 
-            #[cfg(feature = "events")]
-            let event_emitter = Arc::new(Mutex::new(EventEmitter::new()));
+        if self.secret_manager.is_none() {
+            let secret_manager = read_manager_builder
+                .as_ref()
+                .and_then(|data| data.secret_manager.clone())
+                .ok_or(crate::Error::MissingParameter("secret_manager"))?;
 
-            return Ok(AccountManager {
-                #[cfg(not(feature = "events"))]
-                accounts: Arc::new(RwLock::new(
-                    accounts
-                        .into_iter()
-                        .map(|a| AccountHandle::new(a, client.clone(), secret_manager.clone(), storage_manager.clone()))
-                        .collect(),
-                )),
-                #[cfg(feature = "events")]
-                accounts: Arc::new(RwLock::new(
-                    accounts
-                        .into_iter()
-                        .map(|a| {
-                            AccountHandle::new(
-                                a,
-                                client.clone(),
-                                secret_manager.clone(),
-                                event_emitter.clone(),
-                                storage_manager.clone(),
-                            )
-                        })
-                        .collect(),
-                )),
-                background_syncing_status: Arc::new(AtomicUsize::new(0)),
-                client_options: Arc::new(RwLock::new(client_options)),
-                coin_type: Arc::new(AtomicU32::new(coin_type)),
-                secret_manager,
-                #[cfg(feature = "events")]
-                event_emitter,
-                storage_options,
-                storage_manager,
-            });
+            // Update self so it gets used and stored again
+            self.secret_manager.replace(secret_manager);
+        }
+
+        if self.coin_type.is_none() {
+            let coin_type = read_manager_builder
+                .and_then(|data| data.coin_type)
+                .ok_or(crate::Error::MissingParameter("coin_type (IOTA: 4218, Shimmer: 4219)"))?;
+
+            // Update self so it gets used and stored again
+            self.coin_type.replace(coin_type);
+        }
+
+        // Store account manager data in storage
+        #[cfg(feature = "storage")]
+        storage_manager.lock().await.save_account_manager_data(&self).await?;
+
+        let client = self
+            .client_options
+            .clone()
+            .ok_or(crate::Error::MissingParameter("client_options"))?
+            .finish()?;
+
+        #[cfg(feature = "events")]
+        let event_emitter = Arc::new(Mutex::new(EventEmitter::new()));
+
+        #[cfg(feature = "storage")]
+        let accounts = storage_manager.lock().await.get_accounts().await.unwrap_or_default();
+        #[cfg(not(feature = "storage"))]
+        let accounts = Vec::new();
+        let mut account_handles: Vec<AccountHandle> = accounts
+            .into_iter()
+            .map(|a| {
+                AccountHandle::new(
+                    a,
+                    client.clone(),
+                    self.secret_manager
+                        .clone()
+                        .expect("secret_manager needs to be provided"),
+                    #[cfg(feature = "events")]
+                    event_emitter.clone(),
+                    #[cfg(feature = "storage")]
+                    storage_manager.clone(),
+                )
+            })
+            .collect::<_>();
+
+        // If the manager builder is not set, it means the user provided it and we need to update the addresses.
+        // In the other case it was loaded from the database and addresses are up to date.
+        if new_provided_client_options {
+            for account in account_handles.iter_mut() {
+                account.update_account_with_new_client(client.clone()).await?;
+            }
         }
 
         Ok(AccountManager {
-            accounts: Arc::new(RwLock::new(Vec::new())),
+            accounts: Arc::new(RwLock::new(account_handles)),
             background_syncing_status: Arc::new(AtomicUsize::new(0)),
             client_options: Arc::new(RwLock::new(
                 self.client_options
@@ -227,9 +234,9 @@ impl AccountManagerBuilder {
                 .secret_manager
                 .ok_or(crate::Error::MissingParameter("secret_manager"))?,
             #[cfg(feature = "events")]
-            event_emitter: Arc::new(Mutex::new(EventEmitter::new())),
+            event_emitter,
             #[cfg(feature = "storage")]
-            storage_options: StorageOptions { ..Default::default() },
+            storage_options,
             #[cfg(feature = "storage")]
             storage_manager,
         })